requirements.txt
selenium>=4.6
netlib==0.17
psycopg2-binary
python-dotenv
